Mover classe pergunta para arquivo proprio
Por questoes de erros de serializacao se viu ne-
cessario mover a classe para um arquivo proprio
com o objetivo de importa-la antes do inicio da
sessao.
